Added __init__.py modules to folders to make them python packages (fixes imports); GUI application integrated with simulator backend runs without error from app.py entrypoint, but does nothing
