+++ gui/netgrid_model.py
@@ -3,21 +3,26 @@
 
 import numpy as np
 
-from model import Model
+from gui.utils import Model
 
 from network.network_grid import NetworkGrid
 from network.routing_cube import RoutingCube
 
 
 class NetGridPresenter(Model):
-    def __init__(self, netgrid:NetworkGrid):
+    def __init__(self, netgrid:NetworkGrid, max_size:int=8):
         self.netgrid = netgrid
+        self.max_size = max_size
         super().__init__()
 
     
     def get_nodes(self) -> np.ndarray[tuple[int,int,int]]:
         node_map = self.netgrid.node_map
-        return np.array(node_map.keys())
+        node_positions = np.array(list(node_map.keys()), dtype=int)
+
+        nodes = np.zeros((self.max_size,self.max_size,self.max_size))
+        nodes[node_positions] = 1
+        return nodes
     
 
     def next_state(self):
