--- gui/observer.py
@@ -1,9 +1,0 @@
-"""
-"""
-
-class Observer:
-    def __init__(self, model):
-        self._model = model
-
-    def update(self):
-        raise NotImplementedError()
